change the loading widget

+++ lib/Model/notes_app_model.dart
@@ -11,7 +11,7 @@ class NotesAppModel extends HiveObject {
   @HiveField(2)
   final String date;
   @HiveField(3)
-  final int color;
+  final double color;
 
   NotesAppModel(
       {required this.title,
